Moving cmdline functionality into own package. Adjusting wgs84.tmpl to produce switch case for all codes.
