+++ internal/cmdline/filevar.go
@@ -1,4 +1,4 @@
-package main
+package cmdline
 
 import (
 	"fmt"
