+++ internal/cmdline/props_opts.go
@@ -1,4 +1,4 @@
-package main
+package cmdline
 
 import (
 	"encoding/json"
