+++ internal/cmdline/props_opts_test.go
@@ -1,4 +1,4 @@
-package main
+package cmdline
 
 import (
 	"bytes"
