+++ internal/cmdline/template_opts.go
@@ -1,4 +1,4 @@
-package main
+package cmdline
 
 import (
 	"errors"
