+++ internal/cmdline/wkt_opts.go
@@ -1,4 +1,4 @@
-package main
+package cmdline
 
 import (
 	"bufio"
